refactor: split commands.rs into commands/ module
Moves the 8 command functions from the flat commands.rs into a
commands/ directory with dedicated submodules (diff, files, plugins,
stats, tasks). Updates lib.rs to reference the full submodule paths
for Tauri's generate_handler! macro compatibility.
