+++ tauri/src/commands/files.rs
@@ -1,0 +1,66 @@
+use crate::journal_reader;
+use crate::models::*;
+
+#[tauri::command]
+pub fn get_subagent_journal(session_id: String, subagent_id: String) -> Vec<JournalEntry> {
+    let projects_dir = match dirs::home_dir() {
+        Some(h) => h.join(".claude").join("projects"),
+        None => return vec![],
+    };
+
+    let entries = match std::fs::read_dir(&projects_dir) {
+        Ok(e) => e,
+        Err(_) => return vec![],
+    };
+
+    for project_entry in entries.flatten() {
+        let jsonl_path = project_entry
+            .path()
+            .join(&session_id)
+            .join("subagents")
+            .join(format!("{}.jsonl", &subagent_id));
+
+        if jsonl_path.exists() {
+            let state = journal_reader::parse_journal(&jsonl_path, 0, None);
+            let mut result = state.entries;
+            for entry in &mut result {
+                entry.session_id = subagent_id.clone();
+            }
+            return result;
+        }
+    }
+
+    vec![]
+}
+
+#[tauri::command]
+pub fn list_project_files(cwd: String) -> Vec<String> {
+    use ignore::WalkBuilder;
+
+    let mut files = Vec::new();
+    let walker = WalkBuilder::new(&cwd)
+        .hidden(true)
+        .git_ignore(true)
+        .git_global(true)
+        .git_exclude(true)
+        .max_depth(Some(12))
+        .build();
+
+    for entry in walker.flatten() {
+        if !entry.file_type().is_some_and(|ft| ft.is_file()) {
+            continue;
+        }
+        if let Ok(rel) = entry.path().strip_prefix(&cwd) {
+            let rel_str = rel.to_string_lossy().replace('\\', "/");
+            if !rel_str.is_empty() {
+                files.push(rel_str.to_string());
+                if files.len() >= 5000 {
+                    break;
+                }
+            }
+        }
+    }
+
+    files.sort();
+    files
+}
