+++ tauri/src/lib.rs
@@ -62,16 +62,16 @@ pub fn run() {
             ipc::session::delete_session,
             ipc::project::create_project,
             ipc::project::list_projects,
-            commands::get_diff,
-            commands::get_file_versions,
-            commands::get_subagent_journal,
-            commands::get_slash_commands,
-            commands::list_project_files,
-            commands::get_tasks,
-            commands::get_claude_usage_stats,
+            commands::diff::get_diff,
+            commands::diff::get_file_versions,
+            commands::files::get_subagent_journal,
+            commands::plugins::get_slash_commands,
+            commands::files::list_project_files,
+            commands::tasks::get_tasks,
+            commands::stats::get_claude_usage_stats,
             ipc::updater::check_update,
             ipc::updater::install_update,
-            commands::get_changelog,
+            commands::stats::get_changelog,
         ])
         .run(tauri::generate_context!())
         .expect("error while running tauri application");
